+++ test/test_markdown_nested_three_block_block.py
@@ -5793,7 +5793,6 @@ list</p>
 
 
 @pytest.mark.gfm
-@pytest.mark.skip
 def test_nested_three_block_max_block_max_block_max_empty():
     """
     Verify that a nesting of block quote, block quote, block quote, with
@@ -5807,7 +5806,6 @@ def test_nested_three_block_max_block_max_block_max_empty():
         "[block-quote(1,4):   :   > ]",
         "[block-quote(1,9)::   >    > ]",
         "[block-quote(1,14)::   >    >    >\n   >    >    > ]",
-        "[BLANK(1,15):]",
         "[BLANK(1,15):]",
         "[para(2,16):]",
         "[text(2,16):list:]",
@@ -5864,7 +5862,6 @@ def test_nested_three_block_max_block_max_block_max_no_bq1():
 
 
 @pytest.mark.gfm
-@pytest.mark.skip
 def test_nested_three_block_max_block_max_block_max_empty_no_bq1():
     """
     Verify that a nesting of block quote, block quote, block quote, with
@@ -5879,7 +5876,6 @@ def test_nested_three_block_max_block_max_block_max_empty_no_bq1():
         "[block-quote(1,4):   :   > ]",
         "[block-quote(1,9)::   >    > ]",
         "[block-quote(1,14)::   >    >    >]",
-        "[BLANK(1,15):]",
         "[BLANK(1,15):]",
         "[end-block-quote:::False]",
         "[end-block-quote:::False]",
@@ -5937,7 +5933,6 @@ def test_nested_three_block_max_block_max_block_max_no_bq2():
 
 
 @pytest.mark.gfm
-@pytest.mark.skip
 def test_nested_three_block_max_block_max_block_max_empty_no_bq2():
     """
     Verify that a nesting of block quote, block quote, block quote, with
@@ -5952,7 +5947,6 @@ def test_nested_three_block_max_block_max_block_max_empty_no_bq2():
         "[block-quote(1,4):   :   > \n   > ]",
         "[block-quote(1,9)::   >    > ]",
         "[block-quote(1,14)::   >    >    >\n   > ]",
-        "[BLANK(1,15):]",
         "[BLANK(1,15):]",
         "[end-block-quote:::True]",
         "[end-block-quote:::True]",
@@ -6010,7 +6004,6 @@ def test_nested_three_block_max_block_max_block_max_no_bq3():
 
 
 @pytest.mark.gfm
-@pytest.mark.skip
 def test_nested_three_block_max_block_max_block_max_empty_no_bq3():
     """
     Verify that a nesting of block quote, block quote, block quote, with
@@ -6025,7 +6018,6 @@ def test_nested_three_block_max_block_max_block_max_empty_no_bq3():
         "[block-quote(1,4):   :   > ]",
         "[block-quote(1,9)::   >    > ]",
         "[block-quote(1,14)::   >    >    >]",
-        "[BLANK(1,15):]",
         "[BLANK(1,15):]",
         "[end-block-quote:::False]",
         "[end-block-quote:::False]",
@@ -6083,7 +6075,6 @@ list</p>
 
 
 @pytest.mark.gfm
-@pytest.mark.skip
 def test_nested_three_block_max_block_max_block_max_empty_no_bq4():
     """
     Verify that a nesting of block quote, block quote, block quote, with
@@ -6098,7 +6089,6 @@ def test_nested_three_block_max_block_max_block_max_empty_no_bq4():
         "[block-quote(1,4):   :   > ]",
         "[block-quote(1,9)::   >    > \n   >    > ]",
         "[block-quote(1,14)::   >    >    >\n   >    > ]",
-        "[BLANK(1,15):]",
         "[BLANK(1,15):]",
         "[end-block-quote:::True]",
         "[icode-block(2,15):    :]",
@@ -6156,7 +6146,6 @@ def test_nested_three_block_max_block_max_block_max_no_bq5():
 
 
 @pytest.mark.gfm
-@pytest.mark.skip
 def test_nested_three_block_max_block_max_block_max_empty_no_bq5():
     """
     Verify that a nesting of block quote, block quote, block quote, with
@@ -6171,7 +6160,6 @@ def test_nested_three_block_max_block_max_block_max_empty_no_bq5():
         "[block-quote(1,4):   :   > ]",
         "[block-quote(1,9)::   >    > ]",
         "[block-quote(1,14)::   >    >    >]",
-        "[BLANK(1,15):]",
         "[BLANK(1,15):]",
         "[end-block-quote:::False]",
         "[end-block-quote:::False]",
@@ -6229,7 +6217,6 @@ list</p>
 
 
 @pytest.mark.gfm
-@pytest.mark.skip
 def test_nested_three_block_max_block_max_block_max_empty_no_bq6():
     """
     Verify that a nesting of block quote, block quote, block quote, with
@@ -6244,7 +6231,6 @@ def test_nested_three_block_max_block_max_block_max_empty_no_bq6():
         "[block-quote(1,4):   :   > \n   > ]",
         "[block-quote(1,9)::   >    > ]",
         "[block-quote(1,14)::   >    >    >\n   > ]",
-        "[BLANK(1,15):]",
         "[BLANK(1,15):]",
         "[end-block-quote:::True]",
         "[end-block-quote:::True]",
@@ -6302,7 +6288,6 @@ list</p>
 
 
 @pytest.mark.gfm
-@pytest.mark.skip
 def test_nested_three_block_max_block_max_block_max_empty_no_bq7():
     """
     Verify that a nesting of block quote, block quote, block quote, with
@@ -6317,7 +6302,6 @@ def test_nested_three_block_max_block_max_block_max_empty_no_bq7():
         "[block-quote(1,4):   :   > ]",
         "[block-quote(1,9)::   >    > ]",
         "[block-quote(1,14)::   >    >    >]",
-        "[BLANK(1,15):]",
         "[BLANK(1,15):]",
         "[end-block-quote:::False]",
         "[end-block-quote:::False]",