+++ test/test_markdown_nested_three_ordered_ordered.py
@@ -877,7 +877,6 @@ item</p>
 
 
 @pytest.mark.gfm
-@pytest.mark.skip
 def test_nested_three_ordered_max_ordered_max_block_max_empty():
     """
     Verify that a nesting of ordered list, ordered list, block quote, with
@@ -889,9 +888,8 @@ def test_nested_three_ordered_max_ordered_max_block_max_empty():
                > item"""
     expected_tokens = [
         "[olist(1,4):.:1:9:   ]",
-        "[olist(1,10):.:1:15:         :\n]",
+        "[olist(1,10):.:1:15:         :]",
         "[block-quote(1,16):               :               >\n               > ]",
-        "[BLANK(1,17):]",
         "[BLANK(1,17):]",
         "[para(2,18):]",
         "[text(2,18):item:]",
@@ -956,7 +954,6 @@ item</p>
 
 
 @pytest.mark.gfm
-@pytest.mark.skip
 def test_nested_three_ordered_max_ordered_max_block_max_empty_no_bq1():
     """
     Verify that a nesting of ordered list, ordered list, block quote, with
@@ -969,9 +966,8 @@ def test_nested_three_ordered_max_ordered_max_block_max_empty_no_bq1():
                  item"""
     expected_tokens = [
         "[olist(1,4):.:1:9:   ]",
-        "[olist(1,10):.:1:15:         :\n               \n]",
+        "[olist(1,10):.:1:15:         :               \n]",
         "[block-quote(1,16):               :               >]",
-        "[BLANK(1,17):]",
         "[BLANK(1,17):]",
         "[end-block-quote:::False]",
         "[para(2,18):  ]",